# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for synchronizing AxisGTD tasks and configurations.",
        "title": "AxisGTD Sync API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@axisgtd.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Checks if the AxisGTD synchronization service is running.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "index"
                ],
                "summary": "Check service status",
                "responses": {
                    "200": {
                        "description": "AxisGTD synchronization service has been run successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/create": {
            "get": {
                "description": "Creates a new UID with a generated name and sets up the axisgtd table.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "id"
                ],
                "summary": "Create a new UID and axisgtd table",
                "responses": {
                    "200": {
                        "description": "Create ID successful! Your ID is {uidName}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/delete/{name}/{time}": {
            "delete": {
                "description": "Deletes a record from the database based on UID name and time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delete"
                ],
                "summary": "Delete a record by UID name and time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The record's time",
                        "name": "time",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/id/{name}": {
            "get": {
                "description": "Retrieves a list of AxisGTD records associated with the given UID name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "id"
                ],
                "summary": "Get AxisGTD records by UID name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.AxisGTDJsonType"
                            }
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a UID and all associated axisgtd records from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "id"
                ],
                "summary": "Delete a UID and associated axisgtd records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UID and associated records deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ids": {
            "get": {
                "description": "Retrieves the count of axisgtd entries associated with each UID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "id"
                ],
                "summary": "Get counts of axisgtd per UID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.IDSType"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/status/{name}": {
            "get": {
                "description": "Updates the status field of a UID to the opposite value.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Toggle the status of a UID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status toggled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "UID not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sync/{name}": {
            "get": {
                "description": "Retrieves the latest AxisGTD record associated with the specified UID name, ordered by time in descending order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Get the latest AxisGTD record by UID name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The latest AxisGTD record",
                        "schema": {
                            "$ref": "#/definitions/api.AxisGTDJsonType"
                        }
                    },
                    "404": {
                        "description": "UID not found or no records available",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new AxisGTD record into the database for the given UID name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Create a new AxisGTD record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AxisGTD record to create",
                        "name": "todo_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AxisGTDType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "UID not found or UID is disabled",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AxisGTDJsonType": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "time": {
                    "type": "integer"
                },
                "todolist": {
                    "type": "string"
                }
            }
        },
        "api.AxisGTDType": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "todolist": {
                    "type": "string"
                },
                "uidname": {
                    "type": "string"
                }
            }
        },
        "api.IDSType": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}